TODOS by userID

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/app/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import { object, string, boolean } from 'yup';
@@ -30,11 +31,17 @@ let postShema = object({
 });
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  console.log(user);
+  if (!user) {
+    return NextResponse.json('No autorizado', { status: 401 });
+  }
+
   try {
     //const body = await request.json();
     // Usando YUP
     const { description, complete } = await postShema.validate(await request.json());
-    const todo = await prisma.todo.create({ data: { description, complete } });
+    const todo = await prisma.todo.create({ data: { description, complete, userId: user.id } });
     return NextResponse.json(todo)
   } catch (error) {
     return NextResponse.json(error, { status: 400 });
@@ -42,8 +49,14 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+
+  if (!user) {
+    return NextResponse.json('No autorizado', { status: 401 });
+  }
+
   try {
-    const deleteTodoComplete = await prisma.todo.deleteMany({ where: { complete: true } });
+    const deleteTodoComplete = await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
 
     return NextResponse.json(deleteTodoComplete)
   } catch (error) {
